Culture/Skills Array
This file contains work culture and job skills arrays. Both arrays are ready to be included in other javaScript files. There may be one error in the culture array, which is "UX/UI". I think this may only be a skill. If it is an error, it also needs to be fixed in the "cultureSort.html" file.